+++ base_attendance/models/__init__.py
@@ -1,0 +1,5 @@
+# -*- coding: utf-8 -*-
+
+from . import res_attendance
+from . import res_partner
+from . import res_config