+++ pos_debt_notebook/report/pos_debt_report.py
@@ -21,9 +21,10 @@ class PosDebtReport(models.Model):
     config_id = fields.Many2one('pos.config', string='POS', readonly=True)
     company_id = fields.Many2one('res.company', string='Company', readonly=True)
     currency_id = fields.Many2one('res.currency', string='Currency', readonly=True)
+    journal_id = fields.Many2one('account.journal', string='Journals', readonly=True)
 
     state = fields.Selection([('open', 'Open'), ('confirm', 'Validated')], readonly=True)
-    credit_product = fields.Boolean('Credit Product', help="Record is registered as Purchasing credit product", readonly=True)
+    credit_product = fields.Boolean(string='Journal Credit Product', help="Record is registered as Purchasing credit product", readonly=True)
     balance = fields.Monetary('Balance', help="Negative value for purchases without money (debt). Positive for credit payments (prepament or payments for debts).", readonly=True)
     product_list = fields.Text('Product List', readonly=True)
 
@@ -49,7 +50,9 @@ class PosDebtReport(models.Model):
                     session.config_id as config_id,
                     o.company_id as company_id,
                     pricelist.currency_id as currency_id,
-                    o.product_list as product_list
+                    o.product_list as product_list,
+
+                    st.journal_id as journal_id
 
                 FROM account_bank_statement_line as st_line
                     LEFT JOIN account_bank_statement st ON (st.id=st_line.statement_id)
@@ -83,7 +86,9 @@ class PosDebtReport(models.Model):
                     session.config_id as config_id,
                     o.company_id as company_id,
                     pricelist.currency_id as currency_id,
-                    o.product_list as product_list
+                    o.product_list as product_list,
+
+                    pt.credit_product as journal_id
 
                 FROM pos_order_line as pos_line
                     LEFT JOIN product_product pp ON (pp.id=pos_line.product_id)
@@ -93,10 +98,10 @@ class PosDebtReport(models.Model):
 
                     LEFT JOIN pos_session session ON (session.id=o.session_id)
                     LEFT JOIN product_pricelist pricelist ON (pricelist.id=o.pricelist_id)
+                    LEFT JOIN account_journal journal ON (journal.id=pt.credit_product)
                 WHERE
-                    pt.credit_product=true
+                    journal.debt=true
                     AND o.state IN ('paid','done')
-
                 )
                 UNION ALL
                 (
@@ -116,21 +121,24 @@ class PosDebtReport(models.Model):
                     NULL::integer as config_id,
                     inv.company_id as company_id,
                     inv.currency_id as currency_id,
-                    '' as product_list
+                    '' as product_list,
+
+                    pt.credit_product as journal_id
 
                 FROM account_invoice_line as inv_line
                     LEFT JOIN product_product pp ON (pp.id=inv_line.product_id)
                     LEFT JOIN product_template pt ON (pt.id=pp.product_tmpl_id)
                     LEFT JOIN account_invoice inv ON (inv.id=inv_line.invoice_id)
+                    LEFT JOIN account_journal journal ON (journal.id=pt.credit_product)
                 WHERE
-                    pt.credit_product=true
+                    journal.debt=true
                     AND inv.state in ('paid')
                 )
                 UNION ALL
                 (
                 SELECT
                     (-2147483647 + record.id) as id,
-                    NULL::integer as order_id,
+                    record.order_id as order_id,
                     NULL::integer as invoice_id,
                     record.id as update_id,
                     record.balance as balance,
@@ -144,7 +152,8 @@ class PosDebtReport(models.Model):
                     NULL::integer as config_id,
                     record.company_id as company_id,
                     record.currency_id as currency_id,
-                    record.note as product_list
+                    record.note as product_list,
+                    record.journal_id as journal_id
 
                 FROM pos_credit_update as record
                 WHERE